requirements.txt
autopep8==1.6.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
distlib==0.3.4
filelock==3.4.2
idna==3.2
Pillow==9.0.0
platformdirs==2.4.1
protobuf==3.19.1
pycodestyle==2.8.0
pycurl==7.44.1
pyparsing==3.0.0b3
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
virtualenv==20.13.0
